refactor: :ambulance: ts -> js refactoring
