+++ db.json
@@ -1,0 +1,10 @@
+{
+	"user": [
+		{
+			"email": "test@woowa.com",
+			"password": "",
+			"nickname": "",
+			"date": ""
+		}
+	]
+}
