+++ package.json
@@ -4,28 +4,21 @@
 	"description": "",
 	"main": "index.js",
 	"scripts": {
-		"dev": "nodemon --exec ts-node -r tsconfig-paths/register src/index.ts",
-		"build": "tsc -p tsconfig.json",
-		"prestart": "rm -rf ./dist && npm run build",
-		"start": "NODE_ENV=production ts-node -- --type-check -r tsconfig-paths/register ./dist/app.js"
+		"dev": "nodemon src/index.js",
+		"start": "NODE_ENV=production node ./dist/app.js"
 	},
 	"author": "",
 	"license": "ISC",
 	"dependencies": {
+		"bcrypt": "^5.0.1",
 		"cookie-parser": "^1.4.6",
 		"express": "^4.18.1",
 		"http-errors": "^2.0.0",
+		"lowdb": "^1.0.0",
+		"morgan": "^1.10.0",
 		"pug": "^3.0.2"
 	},
 	"devDependencies": {
-		"@types/cookie-parser": "^1.4.3",
-		"@types/express": "^4.17.13",
-		"@types/http-errors": "^1.8.2",
-		"@types/node": "^18.0.1",
-		"nodemon": "^2.0.18",
-		"ts-node": "^10.8.2",
-		"tsc-watch": "^5.0.3",
-		"tsconfig-paths": "^4.0.0",
-		"typescript": "^4.7.4"
+		"nodemon": "^2.0.19"
 	}
 }
