+++ src/index.js
@@ -1,4 +1,4 @@
-import app from './app';
+const app = require('./app');
 
 app.listen(3000, () => {
 	console.log('실행');
