--- src/routes/index.ts
@@ -1,5 +1,0 @@
-import mainRouter from './main';
-import loginRouter from './login';
-import signupRouter from './signup';
-
-export { mainRouter, loginRouter, signupRouter };
