+++ src/routes/login.js
@@ -1,9 +1,8 @@
-import { Router } from 'express';
-
+const { Router } = require('express');
 const router = Router();
 
 router.get('/', (req, res, next) => {
 	res.render('login', { title: 'Login', message: 'test message' });
 });
 
-export default router;
+module.exports = router;
