+++ src/routes/main.js
@@ -1,9 +1,8 @@
-import { Router } from 'express';
-
+const { Router } = require('express');
 const router = Router();
 
 router.get('/', (req, res, next) => {
 	res.render('index', {});
 });
 
-export default router;
+module.exports = router;
