+++ src/utils/asyncHandler.js
@@ -1,0 +1,10 @@
+const asyncHandler = (handler) => {
+	return async (req, res, next) => {
+		try {
+			await handler(req, res, next);
+		} catch (e) {
+			next(e);
+		}
+	};
+};
+module.exports = asyncHandler;
