+++ src/utils/errorHandler.js
@@ -1,0 +1,9 @@
+const errorHandler = (err, req, res, next) => {
+	res.locals.message = err.message;
+	res.locals.error = req.app.get('env') === 'development' ? err : {};
+
+	res.status(err.status || 500);
+	res.render('error');
+};
+
+module.exports = errorHandler;
