--- src/utils/errorHandler.ts
@@ -1,12 +1,0 @@
-import { ErrorRequestHandler } from 'express';
-
-const errorHandler: ErrorRequestHandler = (err, req, res, next) => {
-	res.locals.message = err.message;
-	res.locals.error = req.app.get('env') === 'development' ? err : {};
-
-	// render the error page
-	res.status(err.status || 500);
-	res.render('error');
-};
-
-export default errorHandler;
