+++ src/views/detail.pug
@@ -19,7 +19,7 @@ head
 
       if email
         .input-box.email-field.invalid
-          input.current-input(type="email" placeholder="이메일을 입력해주세요." name="email")
+          input.current-input.email-input(type="email" placeholder="이메일을 입력해주세요." name="email")
           span.success
           button.check-btn 중복확인
       else
@@ -27,10 +27,11 @@ head
           input.current-input(type="email" placeholder="이메일을 입력해주세요." name="email")
           span.success
           button.check-btn 중복확인
-        
-      if email
-        p.message.server-error 중복된 이메일입니다!
-        p.message 유효한 이메일이 아닙니다!
+
+      p.message 유효한 이메일이 아닙니다!
+
+      if email 
+        p.message.error 중복된 이메일입니다!
 
     section.detail-container-section
       .detail-container-item
