+++ src/views/login.pug
@@ -5,10 +5,10 @@ head
   link(rel='stylesheet', href='/static/stylesheets/style.css', type='text/css')
   link(rel='stylesheet', href='/static/stylesheets/login.css', type='text/css')
 
-.app
+#app
   header.header
-    a.header-prev(href="/")
-      img.close(src="/static/images/close2.png")
+    a.prev(href="/")
+      img.close(src="/static/images/cancel.svg")
     span.bold.heading 로그인
 
   form.login-container
