--- tsconfig.json
@@ -1,18 +1,0 @@
-{
-	"extends": "./tsconfig.paths.json",
-	"compilerOptions": {
-		"target": "es5",
-		"module": "Commonjs",
-		"strict": true,
-		"esModuleInterop": true,
-		"skipLibCheck": true,
-		"isolatedModules": true,
-		"forceConsistentCasingInFileNames": true,
-		"allowJs": true,
-		"sourceMap": true,
-		"outDir": "./dist",
-		"typeRoots": ["./node_modules/@types"]
-	},
-	"include": ["./src/*.ts"],
-	"exclude": ["node_modules"]
-}
