--- tsconfig.paths.json
@@ -1,9 +1,0 @@
-{
-	"compilerOptions": {
-		"baseUrl": ".",
-		"paths": {
-			"@src/*": ["src/*"],
-			"@utils/*": ["src/utils/*"]
-		}
-	}
-}
